Introduce OrderedKey and OrderedKeyFunction
An OrderedKey[PK,K] is a proof that there exists a monotone function
from K to PK, which of course requires orderings on K and PK. For
convenience in our particular use-case, we demand class tags for K and
PK.

An OrderedKeyFunction[K1,PK1,K2,PK2] transforms from one pair to
another. In particular, the transformation must be monotonic at the K
and PK levels. Moreover, these transformations must commute with the
respective project methods of OrderedKey[PK,K].