Fix bug in Genotype.gtFromLinear
We had a destructive bug in this function that
caused dosages to never result in a HomRef call.
Our tests were inadequate because we used this
function both in the generator and import, i.e.
testing it against itself.  I added a unit test for
this method.

Fixes #714

Also fixed a bug in OrderedLeftJoinRDD which
was revealed by the Gen.partition update