Fix PipelineSuite file io race
If you follow the call path of a `org.apache.hadoop.fs.LocalFileSystem.mkdirs`,
you'll find that it's actually implemented on
`org.apache.hadoop.fs.FilterFileSystem` which implements it in terms of some
internal `fs` variable, which in `LocalFileSystem`'s case is
`org.apache.hadoop.fs.RawLocalFileSystem`. This class has `mkOneDirWithMode`,
which delegates to `java.io.File`. Now we're at ground truth. This is a well
 documented Java API. It returns true if the directory was created, false
otherwise. No `IOException`s. Going back up through the callstack, we find that
`IOException`s are thrown for a variety of unusual circumstances (like if
you're creating `/foo/bar/baz` and `/foo/bar` is a file that isn't a
directory), but, ultimately, if the directory *already exists* `mkdirs`
returns `false`, it does *not* throw an `IOException`.