Extract CM into asm4s
Now the CM carries around an opaque `AnyRef` state value. It probably
shouldn't even carry this around, but this works for now. The extra
state value will be completely eliminated anyway in a follow up pull
request.

All references to `Type` in the CM were converted to `ClassTag`, which
is the only information we actually need.