Fix: TakeBy Merge Bug, TakeBy Result Not Ordered (#1761)
* bump breeze version

This version of breeze fixes https://github.com/scalanlp/breeze/issues/356
which concerns over-zealous symmetry checking. This issue
was triggered by Hail for at least one of our users:
http://discuss.hail.is/t/fresh-install-fails-eigsymdsuite-and-linearmixedregressionsuite-tests/169/4

* compare takeBy to sortBy

* equality only matters up to irrelevant disorderings

* fix bug in test

* fix test again

* fix TakeByAggregator merge bug

Instead of using the computed sort-keys we were recomputing
the keys when merging. At merge-time, the environment is not
appropriately configured so the results are non-sense.

* tests check for prefixes

* ensure output of TakeBy is ordered

* Clean up TakeBy test a bit