Implicitly verify biallelic.
Implicitly verify biallelic on non-split VDSes on the python side for
methods that require it through the @require_biallelic decorator.

Scala interface now `require`s biallelic expectations.  Other
requirements should be written in this form, too (like TSampleString).
This is based on the decision not to support a user-level Scala
interface that mirror the Python.  We've already left that with, say,
history.