There was a bug converting Call to/from Java, in this line:
        if annotation:

A call is represented by a 0 integer which is False in Python.  I
changed all of these to `if annotation is not None`.

Also, there was a bit of confusion in the call interface inherited
from genotype.  A not call is just a missing Call value (in python)
and a null Call = java.lang.Integer (in Java).