Use array instead of off-heap memory in RegionValue.
Allocating memory off-heap but relying on the GC to deallocate will
never work.  Since regions are allocated for iterators, it is
difficult to know when we can free the memory.  When hasNext returns
false, but ending iteration early will lead to a memory lead.  We
could define something like a CloseableIterator which closes itself at
the end of iteration and clients have to close if they exit early, but
this won't play nicely with Spark (or Scala iterators, for that
matter).

This doesn't matter in the current setup but will make it somewhat
more difficult to push things into C/C++.