Fix bug inserting into missing structs. (#2473)
* Fix bug inserting into missing structs.

When instantiating structs inserting into missing structs, make sure
to set all fields (except the one we're inserting into) to missing.

* Fix python Call handle bug (and probably another annotations) (#2474)

* There was a bug converting Call to/from Java, in this line:

        if annotation:

A call is represented by a 0 integer which is False in Python.  I
changed all of these to `if annotation is not None`.

Also, there was a bit of confusion in the call interface inherited
from genotype.  A not call is just a missing Call value (in python)
and a null Call = java.lang.Integer (in Java).

* Addressed comments.

* Fix region value builder copy bug introduced by stripping required field. (#2475)