Use array instead of off-heap memory in RegionValue. (#2476)
* Use array instead of off-heap memory in RegionValue.

Allocating memory off-heap but relying on the GC to deallocate will
never work.  Since regions are allocated for iterators, it is
difficult to know when we can free the memory.  When hasNext returns
false, but ending iteration early will lead to a memory lead.  We
could define something like a CloseableIterator which closes itself at
the end of iteration and clients have to close if they exit early, but
this won't play nicely with Spark (or Scala iterators, for that
matter).

This doesn't matter in the current setup but will make it somewhat
more difficult to push things into C/C++.

* Minor expr optimizations. (#2477)

* Minor expr optimizations.

Minor, maybe, but, with along with some other changes, made a
significant speedup to a pipline of Konrad's.

* Added some tests.

* Fixed compile error.

* Handle null in Annotation.copy. (#2478)

* Added generic filter_alleles, annotate_alleles. (#2481)

* Added generic filter_alleles, annotate_alleles.

Fixed typo.

switching to generic

Fixed bug.

Fixed a few bugs.

Fix two rv bugs.

Fixed crash.

MemoryBuffer.addUnsafe* requires types agree exactly (including
requireness) since it copies bits.

FunctionRegistry speed improvements.

Convert Call to Code.
Avoid creating Vector (e.g. filtered range).

Remove assert.

Hopefully speed array map, filter.

Use array instead of off-heap memory in RegionValue.

Allocating memory off-heap but relying on the GC to deallocate will
never work.  Since regions are allocated for iterators, it is
difficult to know when we can free the memory.  When hasNext returns
false, but ending iteration early will lead to a memory lead.  We
could define something like a CloseableIterator which closes itself at
the end of iteration and clients have to close if they exit early, but
this won't play nicely with Spark (or Scala iterators, for that
matter).

This doesn't matter in the current setup but will make it somewhat
more difficult to push things into C/C++.

* Addressed comments.

* Some improvements for VEP. (#2483)

* Some improvements for VEP.

RV-ify the easy case.
Preserve partitioning where possible in VSM legacy constructors.

* Fixed bug.