implement a map-like operation on IR
If we view IR as the least fixpoint of an algebra wherein recursive
references are replaced with a type parameter, a la:

IR[T] ::= I32() ...
       |  Cast(T, Type)
       |  ...
       |  Let(T, String, T, Type)
       |  ...

Then IR is a functor and has a natural map operation.

Recur is inspired by this intuition, but in the restricted situation of
Scala's built-in type recursion.