Split CodeNullable from CodeObject (#2609)
* Split CodeNullable from CodeObject

The AnyRef bound is necessary for the ability to reflectively
invoke methods. The Null bound is necessary for operations that
return null as a value of that type.

Unfortunately AnyRef does not imply a Null lower bound. This
PR resolves this issue by formally divorcing the two notions.
Most types will staisfy both bounds enabling the use of
either implicit. I actually am not aware of what types do not
satisfy both bounds.

* fix