force a broader type, appeasing Scala's type system
The type of this RDD is rather complicated because we prepend a String to
an RDD of a universally quantified variable.

`writeTextTable` works with any subtypes of AnyRef because it only requires
that the object defines . Ergo, we don't actually need to know
anything about the type parameter of the RDD. The explicit annotation
that I've added prevents Scala from inferring a complex type that cannot
be expressed without `toString`. Such types trigger a warning from the
Scala compiler.