rv-ified concordance.
I changed the semantics slightly.  Instead of using outerJoinDistinct,
since it really seems like the left and right sides sould be treated
symmetricly, I used an "OrderedZipJoin".  This is like a sort and then
zip, except (1) only the same rows with the same key are zipped
together, and (2) if there are different number of rows with a given
key on the left and right, the side with the fewer rows is padding
with null rows.