make Apply and ApplySpecial work with ExtractAggregators
ExtractAggregators uses a `Ref` with a null type (which is set before
returning to its callee) to avoid traversing the IR tree twice because
the type of the aggregation result is not known until all
the aggregations have been seen.

A non lazy call to `typ` will see the null type (and blow up with an
NPE) when an `ApplySpecial`/`Apply` IR node is copied to replace a
leaf IR aggregation node with a `Ref` to the aggregation result
struct.