Teach IR to ArraySum (#3633)
* Teach IR to ArraySum

This teaches the IR how to compile `hl.agg.array_sum`, which it
previously could not compile. This is the first aggregator whose
input is a non-primitive value. I decided to generalize the seqOp
so that it accepts a Region argument and so that the element is
an IR "intermediate". An IR "intermediate" is a primitive value
or a `Long` offset for values of fundamental type Array or Binary.

* condense lines

* simplify string interpolation