Use class loader from asm4s instead of System
This lets me run tests in SBT. SBT sets up a class loader that it
uses to load freshly compiled test clases and execute tests. This
makes the code-compile-test loop less time consuming.

I will check it against a cluster to ensure that it does not
introduce new class loader issues when code is shipped.