Fix the If unification problem (#4265)
* Fix the If unification problem

The fix is somewhat subtle and relies on PruneDeadFields, which is called (a) by the optimizer and (b) by the compiler, so this is safe. The important piece is that the process of upcasting strips requiredness.

This isn't a great design, but I'm comfortable with it for now since I hope physical types will solve this in The Right Way™ before 0.2 release anyway.

fixes #4134

* address comments

* Added In() to ensure non-optimization

* Add void is of type void

* Test structs