[ci] Recover from missing deploy jobs as well
#4659 teaches CI to recover from a build job gone missing, but
I neglected to teach CI how to recover from a deploy job gone
missing. This follows the same strategy but for deploy jobs.

If a deploy job is not found in the list of refreshed jobs
it is simply removed from the deploy_jobs map. The next heal
stage of CI will kick off a new batch job for whatever the
latest undeployed SHA is.