PTypes 33: Abstract file specs further
I've added a 'compatibility' package in scala. This package includes
case classes used for JSON serialization and deserialization. I also
noticed a design issue with RVDSpec -- there's no file format there.
Do we want to version RVDSpec and RelationalSpec separately, or together?
I think it'll be easier to have one file version for both, and that's
what I've done here.

The next PR will be to create a placeholder EncodedType