Allow outputBufferPos == outputBufferSize in read
This allows seeking to the end of a block or the end of file. It is less
efficient to seek to the end of a block than it is to seek to the start
of the next block as in the case of seeking to the end of a block, both
it and the next block will be decompressed.