Add ability to seek to BGzipInput stream (#4781)
* Start virtual seek implementation for bgz stream

* seek finished?

* Add test, fixup issue test discovered

* Add fixme

* Forgot to commit

* Add invalid seek case

* Address some comments

* Update htsjdk dependency

* Allow outputBufferPos == outputBufferSize in read

This allows seeking to the end of a block or the end of file. It is less
efficient to seek to the end of a block than it is to seek to the start
of the next block as in the case of seeking to the end of a block, both
it and the next block will be decompressed.

* Fix up test cases to be in line with suggestions

* Add getVirtualOffset to BGzipInputStream

* Clean up

* Small fixes

* Address comments

* Minor comment and style edits

* Add a few more tests

* Make virtual offsets unique for a given file

It is no longer allowed to seek to the end of a block (with the
exception of end of file), and as a consequence every position within
the stream has a unique virtual offset.

* Add readRepeatedly to RichInputStream

The new functions will continue reading until either enough bytes are
read into the buffer, or read returns 0 or 1. In either case, the number
of bytes read into the buffer is returned.

* Add check for end of block to getVirtualOffset

This ensures that the offset returned will always be a valid input to
seek, even if the stream has been read to exactly the end of a block.

* Move informational comment

* Add return type annotation to readFully

* Remove currentBlockPos, partially revert 611f23f97