C++ was failing on ubuntu 18.04/gcc 7.3 (#4851)
* C++ was failing on ubuntu 18.04/gcc 7.3

the compiled modules weren't picking up lz4 because the invocation
order was wrong.

proper fix is to pass in a list of libraries, but I think linking
against lz4 unconditionally is OK for the moment.

failed dlopen (and a few other things) should get returned in the
native status, but I just left in the fprintf because it is better
than nothing.

* added darwin prebuilt