PTypes 50: Add PType.canonical constructors
When we remove requiredness from virtual types, requiredness will
live on in both encoded types and physical types. This will require
us to use these types to generate the canonical physical type without
extra missing bits. We will need to use this pattern in, for instance,
TableKeyBy, which may need to encode/decode to shuffle, and will
produce a constructable physical type to decode from.