Move entries array to end of rvRow for MatrixUnionRows
This is a workaround to prevent issues with MatrixUnionRows when the
entries arrays are in different places in the rvRowType in each of the
children.  Furthermore, it prevents issues if the entries array is
pruned and then re-added later in rebuild, where it will often be
inserted, likely by MatrixMapRows, at the end of the rvRowType. This
rearrangement caused the type equality assertion in Optimize to fail.