Move entries array to end of rvRowType for MatrixUnionRows (#4891)
* Add test based on minimal example

* Move entries array to end of rvRow for MatrixUnionRows

This is a workaround to prevent issues with MatrixUnionRows when the
entries arrays are in different places in the rvRowType in each of the
children.  Furthermore, it prevents issues if the entries array is
pruned and then re-added later in rebuild, where it will often be
inserted, likely by MatrixMapRows, at the end of the rvRowType. This
rearrangement caused the type equality assertion in Optimize to fail.

* Update Pretty after MatrixUnionRows changes

    1. MatrixUnionRows needs a new pretty rule to parse properly after
    the previous change.
    2. There was a bug where the case for InsertFields was not wrapping
    identifiers that needed quotations like MatrixType.entriesIdentfier
    causing parse errors.

These issues are now resolved.

* Restructure

* Make type checks match python