[hail][site] fix 404 page (#4953)
Any page that was not a root page did not render properly because
it pointed to relative locations for the css and js resources.
Moreover, the 404 page incorrectly used a relative load for the
navbar. These changes change the template.xslt to use an absolute
protocol-agnostic (useful for testing locally without TLS/SSL) URL
and change 404.xslt to use an absolute load of the navbar.

Currently, 404 is the only page that might appear at a non-root URL.