PTypes 53: Bring back InferType alongside InferPType. All IRs inferred. (#5053)
In last year's PTypes season finale (#4900) I changed Infer to infer the
physical type, and used that to determine the virtual type. This was a
mistake, since there will be plenty of cases where we know the type but
not the physical type (e.g. we have a Ref node). There are also some
nodes where physical types don't even make sense: Literal, TableCollect,
and anything else that fails `CanEmit`.

This PTypes episode changes Infer to InferPType, brings back the old
Infer as InferType, and makes some housekeeping changes to the IR file.