Copy AD and PL to remove requiredness
This was found when testing some of the earlier commits in this series.
If I tried to join a combine_gvcfs output and a single sample mt
(produced with transform_one) it would fail with a type error in
TableMultiWayZipJoin. Adding extra assertions showed that LAD and LPL
were unrequired after a merge.