Change local allele numbering (#5347)
* Change local allele numbering

The numbers used in local alleles now correspond to global allele
indexes rather than global alt-allele indexes

Example, obtaining the allele for LGT[i]:
    Old:
        allele = alleles[LA[LGT[i] - 1] + 1]
    New:
        allele = alleles[LA[LGT[i] - 1]]

* Correct initial LA creation for new numbering

* Use bind in combiner min_rep function

* Copy AD and PL to remove requiredness

This was found when testing some of the earlier commits in this series.
If I tried to join a combine_gvcfs output and a single sample mt
(produced with transform_one) it would fail with a type error in
TableMultiWayZipJoin. Adding extra assertions showed that LAD and LPL
were unrequired after a merge.