Add reannotate function to combiner (#5404)
This joins the output of the combiner with the output of GATK's
GnarlyGenotyper tool. It is intended that this function be run with some
combiner output, and GnarlyGenotyper's output from a sites only VCF
produced after running the summarize function also present in the
vcf_combiner module.

It also adds a little documentation.