[bugfix] Fix persisting of vep, logistic regression, poisson regression (#5416)
None of these methods were appropriately persisting, since the call was
in Scala. Since the Python IR was evaluated each time, the a new RVD
would be created and persisted on each call to backend.execute, and
the previous persisted RVDs would only be used once.

We need to be on the lookout for this pattern as we convert more
operations to MatrixToTableFunctions and friends