[ci2] update status before attempting merge (#6133)
* [ci2] update status before attempting merge

Resolves #6120.

CI is the authority on PR status. Before attempting a merge we should
tell GitHub what we know. We need only update it on the one status
we want to merge, but this code already exists for updating everyone
and it only makes network requests for out of date statuses. This seems
simple and correct.

* bump