[hail] Extend AggArrayPerElement to support a "knownLength" argument. (#6125)
* [hail] Extend AggArrayPerElement to support a "knownLength" argument.

This is required to implement lowering for MatrixMapCols (coming after
a dependent PR goes in). We lower the per-column aggregations to an
`AggArrayPerElement`, but if there are no rows, the result would be
missing.

This PR will allow us to feed in the known number of columns as the
base case.

* add type hints

* fix serialization