[deployment] Make docs deployment logic consistent.
The website will get updated when we deploy a PyPI version. This seems
like the best approach for now -- we don't want new functions hanging
around that people can't use. That's caused trouble in the past.

Also, it seems that from ci to ci2, we've started deploying the docs
by full revision, rather than short revision. The makefile reflects
that.