[hail] Add table and matrix table indices
This currently is exposed in `read_table` and `read_matrix_table` as
undocumented optional parameters `_intervals` and `_filter_intervals`
which takes a list of python `Interval`s that are used either as
a filter or a repartition.

This adds an `IndexedRVDSpec` as the primary container format for
indexed data, and increments the file version to 1.1.0.

One index file is written per partition. For matrix tables, the offset
to a particular key for the entries is stored in the `entries_offset`
field of the annotation that an index may contain. We use the new
`IndexSpec` to retrive the appropriate offset from the index so that we
can seek to the proper offset in the partition.

When writing data with a blocked spec (like the default) we use virtual
index offsets similar to tabix. The high 48 bits are used to indicate
the file offset to the start of a block, and the low 16 bits are used to
indicate the offset from the start of the (possibly decompressed) block.