[batch] resolve ui pages with namespace
If batch is not being run from the default namespace, the url should have the namespace prepended to it. Example `http://localhost:8080/jigold/batches/1/jobs/1`

The jinja template also needs to have the namespace prepended if not default. aiohttp_jinja2 provides this functionality with the built-in `url` function.

Two documentation links:
https://docs.aiohttp.org/en/stable/web_advanced.html#nested-applications
https://aiohttp-jinja2.readthedocs.io/en/stable/#default-globals