[hail] fix my mistake in make
Only compile C/C++ when the user asks for HAIL_COMPILE_NATIVES,
otherwise do not even bother compiling it.

Previously, we always compiled the C code when shadowJar was run,
but we never used said code. Recently, we stopped doing that, but
I failed to account for it when I modified the build.gradle and
Makefile.