[hail] fewer allocations in codecs
The `BlockingInputBuffer` allocates a somewhat large array of
bytes each time it is allocated. As such, it is important to avoid
allocating a `BlockingInputBuffer` for each row if each row is
significantly smaller than the buffer size.

This change removes problematic methods from `RegionValue`, `RVD`,
and `CodecSpec` that have poor performance. In every case, a small
code change enables one allocation per-partition. This required the
implementation of `RestartableByteArrayInputStream` which is a thread-
unsafe version of `ByteArrayInputStream` that, crucially, can
be restarted with a new `Array[Byte]`.

fix