Add ETypes
ETypes are a new abstraction for Serialization and Deserialization of
hail physical values. ETypes have methods for building encoders and
decoders. Every EType needs to know how to recive a value of
a compatible PType and serialize it to an OutputBuffer and to
deserialize data coming from an InputBuffer to a compatible PType.

ETypes are a subset of hail's type system corresponding only to
fundamental serialized types. It is on the PType to maintain the correct
invariants so it's physical structure can be properly decoded.

Every recursion for encoding/decoding nested types like arrays and
structs generates a new method. The hope is to keep methods small so
they can be more easily optimized, and we don't run into method size
issues as often.

No caching is done with the methods currently, and so the following type
will generate three different encode methods for PInt32:
    PTuple[PInt32, PInt32, PString, PInt32]