Encoded Types (#7060)
* Add ETypes

ETypes are a new abstraction for Serialization and Deserialization of
hail physical values. ETypes have methods for building encoders and
decoders. Every EType needs to know how to recive a value of
a compatible PType and serialize it to an OutputBuffer and to
deserialize data coming from an InputBuffer to a compatible PType.

ETypes are a subset of hail's type system corresponding only to
fundamental serialized types. It is on the PType to maintain the correct
invariants so it's physical structure can be properly decoded.

Every recursion for encoding/decoding nested types like arrays and
structs generates a new method. The hope is to keep methods small so
they can be more easily optimized, and we don't run into method size
issues as often.

No caching is done with the methods currently, and so the following type
will generate three different encode methods for PInt32:
    PTuple[PInt32, PInt32, PString, PInt32]

* Add asIdent method to PType

The asIdent method should output a string that can be used to identify
the type while still conforming to JVM identifier rules.

* Switch PackCodecSpec to ETypes

* Remove EmitPack(En|De)coder

All serialization should go through ETypes.

* Address basic usage and style comments

* Add generic getMissingness method for Struct-like types

In order to do this, add Requiredness trait that all types implement.
Then create the appropriate static method.

* Add identBase for structs/tuples

* cleanup

* Check method names for proper identifier characters

* _compatible default implementation error message

* Unify EStruct and ETuple

* Also generate PTuples

* Fix restructuring of PBaseStruct

Can't reference uninitialized fields defined in subclasses, during
initialization, so do lazy initialization of those fields instead.

Also size can be a def.

* bump