[hailctl] support specifying tags for the leader node
This is necessary to most easily support Broad GCP security best practices.
When a project is created a series of networks and firewall rules are
created. Hail needs to set a tag on the leader node that allows
access to the SSH port. This option enables that to be
incorporated into the standard hailctl invocation for a user