[hail/ptypes] Copy IRs before inferring
The `assert(_ptype2 == null)` check in InferPType is breaking here on
certain complex pipelines in a way I don't want to debug.

There's no IR sharing within the IR (see utility I added), but there
must be subtrees that are inferred multiple times in different Compile
calls.

This is a safe stop-gap.