[hail] Inline ApplyIR in compile lowering pass. (#8067)
* [hail] Inline ApplyIR in compile lowering pass.

We're currently emitting the explicit node (without optimization!).
This design is incrementally better, and lets us do ptyping more easily.

The right solution is to do generate a method as the node suggests, but
there are some issues to sort out here, like how to return a missing
value. We may need to return a (possibly null) pointer to an allocated
value, which could be inefficient. Pushing ptypes/requiredness fully
through the system would let us avoid this in many cases.

* oops