[query] Use RDD.coalesce and custom coalescer for non-shuffle repartitioning
Both coalesce and naiveCoalese couldn't handle empty keys because the
partition bounds we generate for unkeyed partitions fail isValid with
allowedOverlap of -1. Even if that was ok (make it so that empty keyed
partition bound sets are automatically valid), we end up duplicating all
the data per partion since RepartitionedOrderedRDD2 uses range bound
queries to determine what data goes into each partition. In the empty
case that duplicates the data once for every partition.

The solution here is to take both coalesce and naiveCoalese away from
using repartition in the shuffle=false case and to using RDD.coalesce.
We need to be sure to use a custom PartitionCoalescer in order to create
identical selection to the previous naiveCoalese that is partitioned
with a key.

fixes #8138