Make index annotation type optional
Because of the nature of index reading, for now, it is one of the last
few uses of PType.canonical and using virtual types to create physical
types. In this particular case, we would write this annotation field as
required and then read it as optional. This worked because of an
accident. `first_annotation` was the last field and its index was not
equal to 0 mod 8, as such it had a missing bit by accident since we make
sure to zero out unused missing bits. So the now optional value looks
present and everything is happy. Once transposed array of structs were
used to encode the data this fell apart as no missing bytes are emitted
for the required structs.

This means, we cannot turn column store on without redoing index specs
to properly use physical types.