[query] Encoder for encoding arrays of structs as 'structs of arrays' (#7803)
* [query] Encoder for encoding arrays of structs as 'structs of arrays'

High level changes are as follows:
+ Add one new EType, ETransposedArrayOfStructs.

The layout of ETransposedArrayOfStructs is as follows:
```
START:
    encoded length (int)
    array's missing bits (sequenceof byte) [if struct not required]
    FOREACH field in struct:
        field's missing bits (sequenceof byte, length is NOTE) [if field not required]
        field elements (sequence of encoded field)

NOTE:
    The length of a field's missing bits is equal to the number of present array elements.
    If the array is of length 100, but only has 10 present elements, then each field's
    missing bits will have 10 bits (2 bytes).
```

* HailContext Flag

* fix compatibility check

* Fix counting

* Fix packed int requiredness, remove unused function

* WIP add all struct represented decoded ptypes to matches

* Remove dead code

* Make index annotation type optional

Because of the nature of index reading, for now, it is one of the last
few uses of PType.canonical and using virtual types to create physical
types. In this particular case, we would write this annotation field as
required and then read it as optional. This worked because of an
accident. `first_annotation` was the last field and its index was not
equal to 0 mod 8, as such it had a missing bit by accident since we make
sure to zero out unused missing bits. So the now optional value looks
present and everything is happy. Once transposed array of structs were
used to encode the data this fell apart as no missing bytes are emitted
for the required structs.

This means, we cannot turn column store on without redoing index specs
to properly use physical types.

* Add flag to explicitly turn off Transposed Array of Struct Encoding

* Revert "Make index annotation type optional"

This reverts commit a7a67ba03fe25967df78c38a7d925a344e1760b3.

* Really don't use transposed array of structs for indicies

* Fix decode compatiblity

* Group decoders into methods

* Fix compile after lir change

* Handle reuse of Code errors as a result of lir

* Address basic typographic comments

+ Rename variables
+ Use matches for checking types
+ Use orEmpty

* Remove etype hint parameter

Solution was to have the test take the PType directly. I believe that
all real world indicies either use the combatiblity IndexSpec, the new
IndexSpec2, or are BGens where we construct the types directly and
properly.

* Address comments

* writeSplitSpecs takes the index specs

Gets around using the HailContext in MatrixMultiWrite

* Miscellaneous ETransposedArrayOfStructs Fixes

+ MethodBuilder => EmitMethodBuilder[_]
+ newField => genFieldThisRef

Nothing serious