[query] Add PCallCode (#8363)
* [query] Add PCallCode

* Rewrite CallStatsAggregator seqOp to use PCode

* Update forEachAllele to use EmitCodeBuilder

* Add scopedCode and emitWithBuilder

These are convenience methods for getting a Code[T] from a CodeBuilder,
and emitting a function from MethodBuilder using CodeBuilder.

* Clean up forEachAllele

It's used in exactly one place. It now takes an EmitCodeBuilder. Like
many CodeBuilder methods/idioms, the closure should refer to the same
CodeBuilder as is passed into the method.