[ci] fix "GitHubAPI object has no attribute 'get'" error
I never tested that PR that got merged (whoops!) and CI tests are insufficient
to catch this case (we should beef those up, asana task added).

The issue was that I thought the method to issue an HTTP get request was `get`,
but it was `getitem`. This PR fixes that. This error occured during `update` and
thus prevented all forward progress of CI.