[query] Fix dosage length check. (#8487)
This illustrates one of the footguns in using CodeBuilder. When using
functionality like ifx, we must be sure to actually put the resultant
code on the CodeBuilder in the block we pass to ifx. Otherwise we won't
add any code and the behavior may be surprising.

This is basically equivalent to storing some Code in a variable and then
never passing that variable into a returned Code, so it is never
executed.