[query] Fix subtle hamming performance regression/bug (#8497)
* [query] Fix subtle hamming performance regression/bug

This change alters the signature of IEmitCode's apply method to take
a thunk returning a PCode, rather than a PCode. This was an issue
because the CodeBuilder is built up in strict append order. In an idiom
like:

    EmitCode.fromI(mb) { cb =>
      cb += setupIsMissing
      val isMissing = missingExpr
      cb += setupThePCode
      val pc = pcodeExpr
      IEmitCode(cb, isMissing, pc)
    }

We were running all of the Code added to the CodeBuilder as a sort of
"setup" before computing the missingness expression. There was no way to
sequence this such that the Code needed to produce the PCode was not
also run in the setup, thus potentially leading to extra work, even when
the result would be missing, like in hamming.

By changing the signature of apply to take a thunk that produces
a PCode, we give callers the ability to control if code is executed
before or after the missingness check.

Co-Authored-By: Patrick Schultz <pschultz@broadinstitute.org>

* bump

Co-authored-by: Patrick Schultz <pschultz@broadinstitute.org>
Co-authored-by: Daniel King <dking@broadinstitute.org>