[many] only the router defines services
There should be one definition of a service. The router should be
the authority on service definitions. The only exceptions are
self-deployed services: gateway, internal-gateway, and router-resolver.

This primarily reduces possibility of error or confusion by removing
duplication. It does not impair hand deploying of any service because
every service (except the self-deployed ones) needs the router anyway.