[docker] automatically build two bioinformatics images
This adds two new Dockerfiles. The first has gsutil and pip-wheel-installed
Hail. The second builds on the first adding a number of bioinformatics tools
that were included in the CCG Tutorial.

Deployment to dockerhub is not trivial because, unfortunately, I need to mount
the docker socket even to download and then upload an image (never starting a
container). I'll design and implement some extension to CI that lets me deploy
images to docker hub later. For now, I used dev deploy to build
these (https://ci.hail.is/batches/33294) and then manually pulled them and
uploaded them to PyPI as hailgenetics/hail:0.2.37 and
hailgenetics/genetics:0.2.37. I particularly appreciate criticism of the names.