[router-resolver] unauthorized also means not developer
Currently, the router-resolver returns 500 if the session id is invalid. Instead,
it should return 401. This collapses two states: not authorized due to not being
a developer and not authorized due to not being logged in. This is unfortunate, but
we should avoid leaking information as to *why* this endpoint is unauthorized to
an attacker. Developers, presumably, are knowledgable enough to figure out why
they cannot log in on their own.