[hailtop.curl] curl should use localized url
Currently, `hailctl curl` uses `external_url` instead of `url`. As a result,
if `hailctl curl` is used inside a GCE VM or on a k8s pod, the url will always
be `....hail.is` to which GCE VMs and k8s pods likely lack credentials.

This was a mistake when I first wrote curl. At that time, I was only using it for
local testing. It will still work for local testing because our deploy configs are
all `external`.