[query/pcode] Add PBetterLocus a new PLocus implementation
Add PBetterLocus is represented as a Long. The high 32 bits are the index of
the contig in its reference genome. The low 32 bits are the position.

A key assumption here is that there will never be more than INT_MAX
contigs in any reference genome. With that, we can simply compare
PBetterLocus with Long comparision as they are all nonnegative.