Bump versions to Spark 2.4.5 and Scala 2.11.12. (#8676)
* Bump versions to Spark 2.4.5 and Scala 2.11.12.

The logic trying to infer the version of various Spark dependencies
was total garbage and almost certainly except for a few specific
cases.  I was feeling aggressive.  I nuked it.  If we want to support
building with other versions of Spark reliably (whcih we don't test)`,
we should find another way.

* bump