[ukbb-rg][admin-pod][router] fix ukbb-rg connectivity
So I both screwed up by including the TLS proxy settings in the location as
well as setting the router to port 80, which the gateway does not respect.
The gateway tries to speak to the router on 443. All the router services
should speak TLS on 443, even if they proxy to a service speaking plain
HTTP on 80.