[hail] teach MemoryBuffer to act like ByteArray(In|Out)putStream (#8731)
* [hail] teach MemoryBuffer to act like ByteArray(In|Out)putStream

The Shuffler uses MemoryBuffer to apply Hail encoders and decoders on byte arrays. For
example, I need to serialize the ETypes into generated bytecode so I can send it to the
Shuffler at run-time. I serialize the EType object to bytes using an encoder and MemoryBuffer
Then, I base64 encode it and store it as a JVM constant pool String. When I need to start
a Shuffle I base64 decode the string, stuff the bytes in a MemoryBuffer and decode it.

* fix formatting