[Query] Continue to replace Code[Unit] with CodeBuilder (#8693)
* [query] Run setup and test missingness when invoking an EmitParam

We previously were just passing the Code in EmitParams to method
invocations without testing missingness. This change fixes this by
wrapping the PCode in a missingness that will return a default value if
the original EmitCode is missing.

* Make TakeByAggregator take EmitParams

Still one test failing when we pass a NULL pointer as a required value.
(wtf)

* Ensure that ALL code is run before emitMethod invocation

Use a CodeBuilder, make sure that we evaluate all code paths, including
a potential EmitCode.m even for required PTypes, just in case there is
code there that has some other side effect.

* Reenable disabled tests

* [query] Disambiguate ClassBuilder and CodeBuilder in aggs

Name the ClassBuilder parameter kb, short klass builder.

* ResultOp to emitI

* Wip on codebuildering aggregators

* Build works

* Compile test scala

* Need to emit all the functions and other fixes

* Fix AppendOnlyBTree foreach Code[Unit] => CodeBuilder translation

* Always add to the codebuilder

* You _have_ to put the code on the codebuilder

* It's missing not present

* Code **MUST** be added to the codebuilder

The key words "MUST", "MUST NOT", "REQUIRED", "SHALL", "SHALL NOT",
"SHOULD", "SHOULD NOT", "RECOMMENDED",  "MAY", and "OPTIONAL" in this
document are to be interpreted as described in RFC 2119.

* Scope variables properly

* Back to parity with old code in DownsampleAggregator

* Add functions for CodeBuilders to invoke methods

Remove previous method for invoking functions with EmitCode return
values. Add an assertion to assist with porting invoke usages over.

* Use CodeBuilder invoke with TakeByAggregator

* Fix usages of EmitParam in invokeCode

* Update some invokes