[site] make namespaced sites work
This fixes site to work inside PRs and dev namespaces. The main fix is to teach
site that, when the namespace is not default, all its resources are located at
`/$NAMESPACE/site`. I also use `subs_filter` to rewrite images, anchors, and
stylesheets that have absolute links (this can be `<a href="/"` or `<a
href="/foo/bar/baz.html"`) to include the namespace prefix.

I also added a test that the website is up and returning 200 with our infrastructure