[hail][ir] clean up IRFunctionRegistry (#8718)
* [hail][ir] clean up IRFunctionRegistry

I rage programmed here a bit. I was pretty frustrated with this class when
trying to debug some other issues in the Shuffler IR. I have two aims with this
PR:
1. Use a clear, consistent naming scheme throughout the file
2. Use concise definitions

In particular, I unified the terminology around types to use these words:
- valueParameterTypes: the type of each value-level parameter
- typeParameters: the type-level parameters, these are always type variables, I
  considered calling them typeVariables, but I like the symmetry with valueParameters
- returnType and returnPType
- typeArguments and valueArguments

I also simplified some anonymous class definitions by providing constructor arguments
instead of methods that are always overridden by `val`s.

Unfortunately, when I got to compiling the tests I discovered we create a function type
that does not exist anywhere in the real code. This did not fit my scheme well and lead
to the duplicative definition of SeededMissingAwareJVMFunction.

Oh, and I changed `IRFunction` to `JVMFunction` because there are already "IR" functions
and an "irRegistry" and it was super confusing to not have the IRFunctions inside the
"irRegsitry". I did not use `CodeFunction` because these are actually implemented by
a number of different JVM Bytecode building tools: `Code`, `PCode`, `EmitCode`, and
`IEmitCode`.

* wip

* fixes

* fixes

* fix

* fixes

* fix