[hail] improve HTML display of relational objects (show)
Our current HTML display of tables flattens the table and concatenates the field
names to produce table headers. This leads to long, unreadable headers.

This change reproduces the nesting structure of the types with several table
header layers. The essential activity is to convert:

```
bing
foo:
  bar:
    baz
    quux
  fizzle:
    fazz:
      fazz1
      fazz2
    fezz
  quork
bang
```

into

```
     foo
     -------------------------------
              fizzle
              ----------------
     bar      fazz
     -------- -----------
bing baz quux fazz1 fazz2 fezz quork bang
```

The bottom layer are the names of the leaves of this tree. Working from the
bottom, a name appears when the row corresponds to that name's tree height. For
this reason `bar` appears lower than `fizzle`. This frustrates finding peer
fields. However, I prefer it. I think I have some sense of visual gravity that
wants bar to fall down.

Anyway, I implement this with some html grunginess in `Table._Show` and a new
class named `PlacementTree`. We construct a `PlacementTree` from a type. It is a
tree whose internal and leaf nodes contain a name, width, and height. It has a
method `to_grid` which converts it to an HTML-table-like structure with "spacer"
elements. Our above example looks like:

```python3
[[(None, 1), ('foo', 6), (None, 1)],
 [(None, 1), (None, 2), ('fizzle', 3), (None, 1), (None, 1)],
 [(None, 1), ('bar', 2), ('fazz', 2), (None, 1) (None, 1), (None, 1)],
 [('bing', 1), ('baz', 1), ('quux', 1), ('fazz1', 1), ('fazz2', 1), ('fezz', 1) ('quork', 1), ('bang', 1)]]
```

The code in `Table._Show` converts this to HTML table rows that looks like:
```html
<tr><td></td><td colspan="6">foo</td><td></td></tr>
<tr><td></td><td colspan="2"></td><td colspan=3>fizzle</td><td></td><td></td></tr>
<tr><td></td><td colspan="2">bar</td><td colspan=2>fazz</td><td></td><td></td><td></td></tr>
<tr><td></td><td>baz</td><td>quux</td><td>fazz1</td><td>fazz2</td><td>fezz</td><td>quork</td><td>bang</td></tr>
```

Which looks like:

<table>
<tr><td></td><td colspan="6">foo</td><td></td></tr>
<tr><td></td><td colspan="2"></td><td colspan=3>fizzle</td><td></td><td></td></tr>
<tr><td></td><td colspan="2">bar</td><td colspan=2>fazz</td><td></td><td></td><td></td></tr>
<tr><td></td><td>baz</td><td>quux</td><td>fazz1</td><td>fazz2</td><td>fezz</td><td>quork</td><td>bang</td></tr>
</table>

---

I also forced cells to not wrap lines. Instead, if they are larger than 500px
they're truncated by the web browser with an ellipsis. Hovering over the cell
displays the full contents in a tool tip.