[batch] timeout for worker activation
This moves `self.activate` into a `wait_for` with a timeout. We use a tight try-except
around the activation code to provide a precise error message if activation fails.

If activation succeeds, we enter the else branch which operates as before. If activation
times out we do not deactivate. This is OK, we probably did not activate. If we did activate,
batch will eventually find out.